LayoutTests/fast/regions/layers/region-zIndex-computedStyle.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../../resources/js-test-pre.js"></script>
        <style>
            #article { -webkit-flow-into: flow; }
            #region { -webkit-flow-from: flow; }
        </style>
    </head>
    <body>
        <p>Test region's computed values for z-index.</p>
        <p>A region establishes a stacking context.</p>
        <p>On success you should see PASS below</p>
        <p id="console"></p>
        <div id="article"></div>
        <div id="region"></div>
    </body>
    <script>
        if (window.testRunner)
            window.testRunner.dumpAsText();
        shouldBeEqualToString('window.getComputedStyle(document.getElementById("region")).zIndex', "0");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
</html>
