safari-600.7-branch/LayoutTests/fast/regions/layers/region-zIndex-computedStyle-expected.txt

Test region's computed values for z-index.

A region establishes a stacking context.

On success you should see PASS below

PASS window.getComputedStyle(document.getElementById("region")).zIndex is "0"
PASS successfullyParsed is true

TEST COMPLETE
